refactor(effect): split window connections and settings loading into own files
